fastapi/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Kubernetes
kubernetes==29.0.0

# Database / ORM
sqlalchemy>=2.0
pymysql

# Validation
pydantic>=2.7
typing-extensions

# JWT
PyJWT

# Password hashing
pwdlib

# OAuth2 form login
python-multipart